ref: parse_config moved to data_coordinator
